+++ modules/apps/transfer/keeper/relay.go
@@ -127,7 +127,13 @@ func (k Keeper) OnRecvPacket(
 		return types.ErrReceiveDisabled
 	}
 
-	receiver, err := sdk.AccAddressFromBech32(data.Receiver)
+	var receiver sdk.AccAddress
+	var err error
+	if k.addressCodec != nil {
+		receiver, err = k.addressCodec.StringToBytes(data.Receiver)
+	} else {
+		receiver, err = sdk.AccAddressFromBech32(data.Receiver)
+	}
 	if err != nil {
 		return errorsmod.Wrapf(ibcerrors.ErrInvalidAddress, "failed to decode receiver address: %s", data.Receiver)
 	}
@@ -195,7 +201,7 @@ func (k Keeper) OnRecvPacket(
 		if err := k.BankKeeper.SendCoins(
 			ctx, moduleAddr, receiver, sdk.NewCoins(voucher),
 		); err != nil {
-			return errorsmod.Wrapf(err, "failed to send coins to receiver %s", receiver.String())
+			return errorsmod.Wrapf(err, "failed to send coins to receiver %s", data.Receiver)
 		}
 
 	}
@@ -253,7 +259,13 @@ func (k Keeper) refundPacketTokens(
 ) error {
 	// NOTE: packet data type already checked in handler.go
 
-	sender, err := sdk.AccAddressFromBech32(data.Sender)
+	var sender sdk.AccAddress
+	var err error
+	if k.addressCodec != nil {
+		sender, err = k.addressCodec.StringToBytes(data.Sender)
+	} else {
+		sender, err = sdk.AccAddressFromBech32(data.Sender)
+	}
 	if err != nil {
 		return err
 	}