+++ CHANGELOG.md
@@ -34,6 +34,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## [Unreleased]
+
+* [\#8573](https://github.com/cosmos/ibc-go/pull/8573) Support custom address codecs in transfer.
+
 ## [v10.3.0](https://github.com/cosmos/ibc-go/releases/tag/v10.3.0) - 2025-06-06
 
 ### Features